+++ src/main/java/com/eiv/repositories/LineaRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.eiv.entities.LineaEntity;
 
-public interface LineaRepository extends JpaRepository<LineaEntity, Integer> {
+public interface LineaRepository extends JpaRepository<LineaEntity, Long> {
 
     @Query("SELECT MAX(e.id) FROM LineaEntity e")
-    public Optional<Integer> getMax();
+    public Optional<Long> getMax();
 }
