+++ src/main/java/com/eiv/repositories/TipoDocumentoRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.eiv.entities.TipoDocumentoEntity;
 
-public interface TipoDocumentoRepository extends JpaRepository<TipoDocumentoEntity, Integer> {
+public interface TipoDocumentoRepository extends JpaRepository<TipoDocumentoEntity, Long> {
 
     @Query("SELECT MAX(e.id) FROM TipoDocumentoEntity e")
-    public Optional<Integer> getMax();
+    public Optional<Long> getMax();
 }
